requirements.txt
fastapi
uvicorn[standard]
pydantic
duckduckgo-search
newspaper3k
httpx
python-dotenv
